UIX: add TaskUtilities and AwaitProvider for easier use of async/await/Tasks in other mods
